+++ packages/cli/src/react/jsx/utils/constants.ts
@@ -6,7 +6,7 @@ export const DECLARE_STATIC_FUNCTION = 'declareStatic';
 export const DERIVE_FUNCTION = 'derive';
 export const MSG_REGISTRATION_FUNCTION = 'msg';
 export const T_REGISTRATION_FUNCTION = 't';
-export const T_GLOBAL_REGISTRATION_FUNCTION = 't';
+export const T_GLOBAL_REGISTRATION_FUNCTION = T_REGISTRATION_FUNCTION;
 export const INLINE_TRANSLATION_HOOK = 'useGT';
 export const INLINE_TRANSLATION_HOOK_ASYNC = 'getGT';
 export const INLINE_MESSAGE_HOOK = 'useMessages';
@@ -18,11 +18,17 @@ export const TRANSLATION_COMPONENT = 'T';
 export const STATIC_COMPONENT = 'Static';
 export const DERIVE_COMPONENT = 'Derive';
 export const BRANCH_COMPONENT = 'Branch';
-export const PLURAL_COMPONENT = 'Plural';
-export const VAR_COMPONENT = 'Var';
 export const DEFAULT_GT_IMPORT_SOURCE = 'gt-react/browser';
 export const INTERNAL_TRANSLATION_COMPONENT = 'GtInternalTranslateJsx';
 export const INTERNAL_VAR_COMPONENT = 'GtInternalVar';
+
+// Variable components
+export const VAR_COMPONENT = 'Var';
+export const DATETIME_COMPONENT = 'DateTime';
+export const RELATIVE_TIME_COMPONENT = 'RelativeTime';
+export const CURRENCY_COMPONENT = 'Currency';
+export const NUM_COMPONENT = 'Num';
+export const PLURAL_COMPONENT = 'Plural';
 
 // GT translation functions
 export const GT_TRANSLATION_FUNCS = [
@@ -38,14 +44,15 @@ export const GT_TRANSLATION_FUNCS = [
   TRANSLATION_COMPONENT,
   STATIC_COMPONENT,
   DERIVE_COMPONENT,
-  'Var',
-  'DateTime',
-  'Currency',
-  'Num',
+  VAR_COMPONENT,
+  DATETIME_COMPONENT,
+  RELATIVE_TIME_COMPONENT,
+  CURRENCY_COMPONENT,
+  NUM_COMPONENT,
   BRANCH_COMPONENT,
-  'Plural',
   INTERNAL_TRANSLATION_COMPONENT,
   INTERNAL_VAR_COMPONENT,
+  PLURAL_COMPONENT,
 ];
 // GT String translation functions
 export const STRING_REGISTRATION_FUNCS = [
@@ -61,10 +68,11 @@ export const GT_DERIVE_STRING_FUNCTIONS = [
 
 // Valid variable components
 export const VARIABLE_COMPONENTS = [
-  'Var',
-  'DateTime',
-  'Currency',
-  'Num',
+  VAR_COMPONENT,
+  DATETIME_COMPONENT,
+  RELATIVE_TIME_COMPONENT,
+  CURRENCY_COMPONENT,
+  NUM_COMPONENT,
   STATIC_COMPONENT,
   DERIVE_COMPONENT,
   INTERNAL_VAR_COMPONENT,