+++ packages/tanstack-start/src/index.ts
@@ -3,7 +3,7 @@ export { GTProvider } from './provider/GTProvider';
 export { getTranslations } from './functions/getTranslations';
 
 // TODO: refactor all component/function exports into an index file
-export { T, Var, LocaleSelector } from 'gt-react';
+export { T, Var, LocaleSelector, RelativeTime } from 'gt-react';
 export { getLocale } from './functions/getLocale';
 export { getVersionId } from './functions/getVersionId';
 export { getGT, getMessages } from 'gt-i18n/internal';