+++ packages/next-lint/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @generaltranslation/gt-next-lint
 
+## 13.0.1
+
+### Patch Changes
+
+- [#1161](https://github.com/generaltranslation/gt/pull/1161) [`eca3d8d`](https://github.com/generaltranslation/gt/commit/eca3d8d8298969258bb4ab576b698c48cfbc318f) Thanks [@moss-bryophyta](https://github.com/moss-bryophyta)! - Update logo blocks in READMEs
+
+- Updated dependencies [[`eca3d8d`](https://github.com/generaltranslation/gt/commit/eca3d8d8298969258bb4ab576b698c48cfbc318f)]:
+  - gt-next@6.15.1
+
+## 13.0.0
+
+### Patch Changes
+
+- Updated dependencies [[`df6bea8`](https://github.com/generaltranslation/gt/commit/df6bea819a4274018d6d99c7d3e00e7c5372ccbc)]:
+  - gt-next@6.15.0
+
 ## 12.0.8
 
 ### Patch Changes