+++ packages/python-extractor/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @generaltranslation/python-extractor
 
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies [[`eca3d8d`](https://github.com/generaltranslation/gt/commit/eca3d8d8298969258bb4ab576b698c48cfbc318f)]:
+  - generaltranslation@8.2.1
+
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [[`9d2349c`](https://github.com/generaltranslation/gt/commit/9d2349cfc41862d9e3d8364659b678055b9fa290), [`df6bea8`](https://github.com/generaltranslation/gt/commit/df6bea819a4274018d6d99c7d3e00e7c5372ccbc)]:
+  - generaltranslation@8.2.0
+
 ## 0.2.3
 
 ### Patch Changes