+++ packages/sanity/CHANGELOG.md
@@ -1,5 +1,28 @@
 # gt-sanity
 
+## 2.0.1
+
+### Patch Changes
+
+- [#1161](https://github.com/generaltranslation/gt/pull/1161) [`eca3d8d`](https://github.com/generaltranslation/gt/commit/eca3d8d8298969258bb4ab576b698c48cfbc318f) Thanks [@moss-bryophyta](https://github.com/moss-bryophyta)! - Update logo blocks in READMEs
+
+- Updated dependencies [[`eca3d8d`](https://github.com/generaltranslation/gt/commit/eca3d8d8298969258bb4ab576b698c48cfbc318f)]:
+  - generaltranslation@8.2.1
+
+## 2.0.0
+
+### Major Changes
+
+- [#1164](https://github.com/generaltranslation/gt/pull/1164) [`2df12cf`](https://github.com/generaltranslation/gt/commit/2df12cf1b28d1ae3fc75222c4aff270ce96a7c28) Thanks [@brian-lou](https://github.com/brian-lou)! - gt-sanity v2: React 19; additional config options
+  See https://generaltranslation.com/en-US/blog/gt-sanity_v2_0_0 for more details.
+
+## 1.1.31
+
+### Patch Changes
+
+- Updated dependencies [[`9d2349c`](https://github.com/generaltranslation/gt/commit/9d2349cfc41862d9e3d8364659b678055b9fa290), [`df6bea8`](https://github.com/generaltranslation/gt/commit/df6bea819a4274018d6d99c7d3e00e7c5372ccbc)]:
+  - generaltranslation@8.2.0
+
 ## 1.1.30
 
 ### Patch Changes